add initial driver implementation
this contains some basic test coverage using `embedded-hal-mock` but
requires dbrgn/embedded-hal-mock#52 to support the `PwmPin` trait. for
the time being the PR branch is being used as a git dependency, this
should be changed back to a normal dependency once the PR has been
merged and a new release of `embedded-hal-mock` has been published.