migrate to `embedded-hal=1.0.0.rc1`
as alpha.10 of embedded-hal has added `SetDutyCycle` as the equivalent /
successor of the old `PwmPin` support for it can now be added here.

since e-h 1.0.0 will be released at the end of this year it's safe to
just directly migrate to the RC1 instead of adding it as an optional
support (requiring another breaking change in a later step which would
then remove e-h 0.2 support).