change `forbid` for warnings & unused to `deny`
a new warning was caused by the `?` operator as it contains
`#[allow(unreachable_code)]`, which clashes with the previous `forbid`
rule.

as `forbid` does not allow `deny` to override it this would cause a
compile error, however due to a compiler bug this was previously not
reported at all and has now been changed to a warning in a recent(-ish)
version of rustc and will eventually become a compile error.

with `deny` it's still possible to use `allow`, thus there's no warning

see also rust-lang/rust#76053

thanks @jannic