remove `defmt` dependency & feature
based on the discussion on GitHub we don't need this dependency. it's
the only 0.x dependency currently present in the crate and doesn't
provide a lot of value since consuming code can use
`defmt::Debug2Format` when needed. the single `defmt::debug!` statement
in `Motor::drive` was not very helpful anyway if two motors were
connected as it did not differentiate between the motors in the message.

since this was the only optional feature there's no need anymore to test
with `--all-features` in the CI. this has been removed from the build
matrix (but the possibility to add it later has been kept as it doesn't
make the build script unnecessarily complicated).